+++ @remirror/editor-wysiwyg/src/wysiwyg-theme.ts
@@ -1,3 +1,5 @@
+
+
 import DefaultStyled, { CreateStyled } from '@emotion/styled';
 import { Cast } from '@remirror/core';
